[refactor][select_group] iterate over dialog list
rather than fetching the entire dialog list altogether in one go
( which also sometimes brought in duplicates ). This is also more easy
on the API and is less prone to `FloodError`s.
The other problem with fetching the entire dialog list was the need to
use multiple loops to get the correct list properly filtered.
This change also brings back the mega-group filter that was previously
removed when function-izing `select_group` from `add_users_to_group` and
adds check for forbidden chats and channels